[admin] [organization] add restore organization

+++ database/migrations/2023_09_27_165511_add_name_to_invitations_table.php
@@ -22,6 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('invitations', function (Blueprint $table) {
+            if (Schema::getConnection()->getName() == 'sqlite') {
+                return;
+            }
             $table->dropColumn('name');
         });
     }
