+++ database/migrations/2023_10_04_173605_add_should_be_invited_at_to_invitations_table.php
@@ -22,6 +22,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('invitations', function (Blueprint $table) {
+            if (Schema::getConnection()->getName() == 'sqlite') {
+                return;
+            }
+
             $table->dropColumn('should_be_invited_at');
         });
     }
